Clone repos at latest tag by default
This sets the git clone pieces to checkout at the latest (by commit date) tag
for the repos by default.  Should also tag a tag as an option for the individual
clone/fetches.  Also adds a few lines to try to get the repos via ssh from the
sot org if possible.