Replace GIT_DESCRIBE_TAG with more template logic
This looks more complicated, but helps to guarantee that package
builder cannot build a package that "masquerades" as a certain version
unless the commit from which it was built was exactly at that tagged
version.  For default builds, where the most recent (by commit) tag
is used as the build, this new template logic changes nothing.  However
if one builds a package using ska3_builder with the --tag option at a commit,
the resulting conda package will have at least a 'dev' string in the package
 version unless the selected commit is on a tag.