Add a post-link.sh to fix shebang rewrite issues
It looks like conda/conda-build at the version we're using replaces
all shebang python scripts with ${PREFIX}/bin/python shebangs.
Unfortunately, that process is just chucking or overwriting the " -E "
option we have in flt_envs to ignore environment variables.
I'm betting this is fixed in a newer version of the conda
tools, but for now this script just does another
shebang rewrite after the template is installed.