Move sherpa to ska3-core
The original idea of keeping sherpa in ska3-flight was that the
sherpa packages themselves are not in the conda default channel but instead
are in our repo.  However, from an os/arch management perspective
it makes more sense to put sherpa in ska3-core so it will only get
pulled in to the build for the operating systems that need it
(and for which we can build packages), not linux32.