Update ruff-base.toml to ignore PLW1641 (#1556)
Rule PLW1641 requires that any class that implements `__eq__` also implements `__hash__`. 

This is misguided since in practice it is very common for classes to define a "liberal" view of equality where objects that compare equal do not have the same hash. As a simple example, `np.dtype("float64") == "float64"` is `True`, while the `hash()` values are not equal.

In our code base it is fine for a class object to not be hashable by default, which is what can happen by ignoring PLW1641.